+++ AvgScript/src/lib/iterateParams.ts
@@ -1,0 +1,27 @@
+import * as vscode from 'vscode';
+import { InlayHintType, commandInfoList } from './dict';
+import { iterateLines } from './iterateLines';
+import { iterateScripts } from './iterateScripts';
+import { getAllParams } from './utilities';
+
+export async function iterateParams(
+    newScriptCallback: (script: string, document: vscode.TextDocument) => void,
+    paramCallback: (initScript: string,
+        script: string, lineNumber: number,
+        currentType: InlayHintType, currentParam: string) => void) {
+    await iterateScripts((script: string, document: vscode.TextDocument) => {
+        newScriptCallback(script, document);
+    },
+        (initScript: string,
+            script: string,
+            line: string, lineNumber: number,
+            lineStart: number, lineEnd: number,
+            firstLineNotComment: number) => { },
+        (initScript: string,
+            script: string, lineNumber: number, line: string,
+            currentType: InlayHintType, currentParam: string) => {
+            paramCallback(initScript,
+                script, lineNumber,
+                currentType, currentParam);
+        });
+}